feat:   配合前端模板工程，vue-admin-wonderful；

+++ kewen-auth-rabc/src/main/java/com/kewen/framework/auth/rabc/controller/UserController.java
@@ -88,7 +88,7 @@ public class UserController {
         PageResult<SysUser> result = new PageResult<>();
         result.setPage(pageReq.getPage());
         result.setSize(pageReq.getSize());
-        result.setTotal((int)page.getTotal());
+        result.setTotal(page.getTotal());
         result.setData(records);
         return Result.success(result);
     }
