+++ kewen-auth-rabc/src/main/java/com/kewen/framework/auth/rabc/model/PageReq.java
@@ -7,8 +7,8 @@ import javax.validation.constraints.NotNull;
 @Data
 public class PageReq {
     @NotNull
-    private Integer page;
+    private Long page;
     @NotNull
-    private Integer size;
+    private Long size;
     private String name;
 }
